+++ packages/platform-nodejs/src/index.ts
@@ -4,6 +4,10 @@ import "source-map-support/register";
 import "reflect-metadata";
 
 import fetch from "node-fetch";
+import WebSocket from 'ws';
 
 if (!globalThis.fetch)
     globalThis.fetch = <typeof globalThis.fetch><unknown>fetch;
+
+if (!globalThis.WebSocket)
+    globalThis.WebSocket = WebSocket;