+++ packages/web-server/src/input.ts
@@ -33,17 +33,31 @@ export interface QueryParamOptions {
  * Apply to a parameter to indicate that it represents a query parameter (ie foo in /bar?foo=1)
  * @param name 
  */
-export function QueryParam(name? : string, options? : QueryParamOptions) {
+ export function QueryParam(name? : string, options? : QueryParamOptions) {
 	return InputAnnotation.decorator({
 		validTargets: [ 'parameter' ],
 		allowMultiple: false
 	})({ 
-		type: 'query', 
+		type: 'queryParam', 
 		name,
 		default: options?.default
 	});
 }
 
+/**
+ * Apply to a parameter to indicate that it represents a query parameter (ie foo in /bar?foo=1)
+ * @param name 
+ */
+ export function QueryParams() {
+	return InputAnnotation.decorator({
+		validTargets: [ 'parameter' ],
+		allowMultiple: false
+	})({
+		type: 'queryParams',
+		name: null,
+		default: {}
+	});
+}
 
 /**
  * Apply to a parameter to indicate that it represents a session parameter (ie foo in /bar?foo=1)